+++ core/src/main/java/com/afs/dragbits/autos/EstadoAuto.java
@@ -1,0 +1,8 @@
+package com.afs.dragbits.autos;
+
+public enum EstadoAuto {
+    ESTATICO,
+    AVANZANDO,
+    CAMBIANDO_MARCHA,
+    NITRO
+}