+++ core/src/main/java/com/afs/dragbits/hud/Semaforo.java
@@ -15,16 +15,6 @@ import com.afs.dragbits.util.SpriteSheetLoader;
  */
 public class Semaforo extends ElementoHUD {
 
-    public enum EstadoSemaforo {
-        APAGADO,      // Frame 0
-        LUZ_1,        // Frame 1
-        LUZ_2,        // Frame 2
-        LUZ_3,        // Frame 3
-        VERDE,        // Frame 4 (Largada)
-        SALIDA_FALSO, // Frame 5 (Infracción)
-        FINALIZADO    // Desaparece de pantalla
-    }
-
     private EstadoSemaforo estadoActual;
     private Texture spriteSheet;
     private TextureRegion[] frames;