+++ core/src/main/java/com/afs/dragbits/menurivales/TipoCarrera.java
@@ -1,0 +1,6 @@
+package com.afs.dragbits.menurivales;
+
+public enum TipoCarrera {
+    LEGAL,
+    ILEGAL
+}